tests(ticdc): move integration into tests
