+++ tests/mq_protocol_tests/framework/mysql/single_table.go
@@ -18,7 +18,7 @@ import (
 	"time"
 
 	"github.com/pingcap/log"
-	"github.com/pingcap/ticdc/integration/framework"
+	"github.com/pingcap/ticdc/tests/mq_protocol_tests/framework"
 )
 
 const (
