+++ tests/mq_protocol_tests/tests/case_alter.go
@@ -18,7 +18,7 @@ import (
 	"math/rand"
 
 	"github.com/pingcap/errors"
-	"github.com/pingcap/ticdc/integration/framework"
+	"github.com/pingcap/ticdc/tests/mq_protocol_tests/framework"
 )
 
 // AlterCase is base impl of test case for alter operation
