+++ tests/mq_protocol_tests/tests/case_many_types.go
@@ -18,10 +18,10 @@ import (
 	"math"
 	"time"
 
-	"github.com/pingcap/ticdc/integration/framework"
-	"github.com/pingcap/ticdc/integration/framework/avro"
-	"github.com/pingcap/ticdc/integration/framework/canal"
-	"github.com/pingcap/ticdc/integration/framework/mysql"
+	"github.com/pingcap/ticdc/tests/mq_protocol_tests/framework"
+	"github.com/pingcap/ticdc/tests/mq_protocol_tests/framework/avro"
+	"github.com/pingcap/ticdc/tests/mq_protocol_tests/framework/canal"
+	"github.com/pingcap/ticdc/tests/mq_protocol_tests/framework/mysql"
 )
 
 // ManyTypesCase is base impl of test case for different types data
